MyStuff/Python/kahbot/requirements.txt
discord.py==1.5.1
discord==1.0.1
python-dotenv==0.15.0